Feat #30 - Templates/Home 생성
